Invoke original inherited callback when subclassed
When AASM is included into a class, it defines inherited, clobbering
the original definition.  This problem becomes readily apparent with STI
in Rails, where the original method is part of the implementation of
inheritable attributes, a feature relied on for several ActiveRecord
features such as callbacks and timestamp recording.

Move the implementation of inherited to ClassMethods and invoke the
original method via super.