Add :after_commit callbacks to events.
Event-specific :after_commit callbacks are called in the order that the events
were called.  They are called even when the record is manually persisted.  For
the sake of consistency, the same is now true of the state :after_commit
callback.  As in that case, this feature is specific to active record.