Added tags file to gitignore. This is the index file created by ctags which is used by some text editors and is developer environment specific.
Added the option to use :if syntax as syntactic sugar for guard.

Added the option to use :unless syntax which essentially behaves
as a negated guard.

Added a couple of tests for the :unless syntax since the logic was
slightly different.

Noticed a couple of tests were just returning true/false at
the end rather than checking for expectations.  Fixed a couple
of those; still some remaining.

Changed AASM#perform method to allowed?