Allow multiple transitions in a single event with the same to and from states
As reported in https://github.com/aasm/aasm/issues/372 and https://github.com/aasm/aasm/issues/362,
if two transitions in the same event have the same `to` and `from` states but differ in their
guards, if any transition is valid, the first transition is the one that's always executed.
This attempts to resolve that issue.