Add current state detail to minitest transition_from assertions
It's useful when reviewing a failure to know which state you did get by firing
the transition, not just the one you expected, as this can help with tracing
down the error in the state machine. This brings the minitest helper closer to
the output from the rspec version which already includes the actual state not
just the expected one.